app/angels/requirements.txt
django-countries>=7.6.1
django-cors-headers>=4.9.0
django-environ>=0.10.0
